Moved most of source code into src module to allow nice relative imports
